% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blender_endmatter.R
\name{add_blender_endmatter}
\alias{add_blender_endmatter}
\title{Add standard ending boilerplate to a Blender rendering script}
\usage{
add_blender_endmatter(script, filepath = tempfile(fileext = ".blend"))
}
\arguments{
\item{script}{The Python script to append the generated code onto.}

\item{filepath}{The file path to save the output file to. Must end with
`.blend`.}
}
\value{
A length 1 character vector containing the Blender Python script with
ending boilerplate added.
}
\description{
Add standard ending boilerplate to a Blender rendering script
}
