% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_camera.R
\name{add_camera}
\alias{add_camera}
\title{Add a camera object to a Blender scene}
\usage{
add_camera(
  script,
  align = c("WORLD", "VIEW", "CURSOR"),
  location = c(0, 0, 0),
  rotation = c(0, 0, 0),
  convert_rotations = TRUE
)
}
\arguments{
\item{script}{The Python script to append the generated code onto.}

\item{align}{Character: alignment of the new camera object. Options include
"WORLD" (align the new object to the world), "VIEW" (align the new object to
the view), and "CURSOR" (use the 3D cursor orientation for the new object).}

\item{location}{A numeric vector with length 3 specifying the x, y, and z
grid coordinates for the new camera.}

\item{rotation}{A numeric vector with length 3 specifying the x, y, and z
rotation for the new camera.}

\item{convert_rotations}{Logical: convert `rotation` to radians? Set to
`FALSE` if `rotation` is already in radians.}
}
\value{
A length 1 character vector containing the Blender Python script with
code to add a camera object added.
}
\description{
Add a camera object to a Blender scene
}
