% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_light.R
\name{add_light}
\alias{add_light}
\title{Add a light object to a Blender scene}
\usage{
add_light(
  script,
  type = c("POINT", "SUN", "SPOT", "AREA"),
  radius = 0,
  align = c("WORLD", "VIEW", "CURSOR"),
  location = c(0, 0, 0),
  rotation = c(0, 0, 0),
  convert_rotations = TRUE,
  energy = 10
)
}
\arguments{
\item{script}{The Python script to append the generated code onto.}

\item{type}{Character: Type of light source to use. Options include "POINT"
(omnidirectional point light source), "SUN" (constant direction parallel ray
light source), "SPOT" (directional cone light source), and "AREA"
(directional area light source).}

\item{radius}{Numeric: Radius of the light source.}

\item{align}{Character: alignment of the new light object. Options include
"WORLD" (align the new object to the world), "VIEW" (align the new object to
the view), and "CURSOR" (use the 3D cursor orientation for the new object).}

\item{location}{A numeric vector with length 3 specifying the x, y, and z
grid coordinates for the new light.}

\item{rotation}{A numeric vector with length 3 specifying the x, y, and z
rotation for the new light.}

\item{convert_rotations}{Logical: convert `rotation` to radians? Set to
`FALSE` if `rotation` is already in radians.}

\item{energy}{Numeric: The power ("wattage") of the light object.}
}
\value{
A length 1 character vector containing the Blender Python script with
code to add a light object added.
}
\description{
Add a light object to a Blender scene
}
