% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_primitive.R
\name{add_primitive}
\alias{add_primitive}
\alias{add_primitive,mvdf_obj-method}
\alias{add_primitive,mvdf_simple_material-method}
\alias{add_mesh_primitive}
\alias{add_surface_primitive}
\alias{add_curve_primitive}
\title{Add code to add primitives to a Blender script}
\usage{
add_primitive(object, script, category, primitive, location = NULL, ...)

\S4method{add_primitive}{mvdf_obj}(
  object,
  script,
  category,
  primitive = "ico_sphere",
  location = NULL,
  ...
)

\S4method{add_primitive}{mvdf_simple_material}(
  object,
  script,
  category,
  primitive = "ico_sphere",
  location = NULL,
  ...
)

add_mesh_primitive(
  script,
  object,
  primitive = "ico_sphere",
  location = NULL,
  ...
)

add_surface_primitive(
  script,
  object,
  primitive = "torus",
  location = NULL,
  ...
)

add_curve_primitive(
  script,
  object,
  primitive = "bezier_circle",
  location = NULL,
  ...
)
}
\arguments{
\item{object}{An object inheriting from [mvdf_obj] which will be used to
calculate the x, y, and z positions for each primitive. One primitive will be
created for each row of `mvdf(object)`.}

\item{script}{The Python script to append the generated code onto.}

\item{category}{What category of primitive to add. This function works well
with `mesh`, `curve`, and `surface` primitives.}

\item{primitive}{The primitive type to create. See the official Blender
documentation at
\url{https://docs.blender.org/api/blender_python_api_current/bpy.ops.html}
for a full list of available primitives. Each category has a different set of
primitives available.}

\item{location}{Either `NULL` (the default) or a vector of strings (in the
format `location=(x, y, z)`) specifying the location for the origin of each
primitive. If `NULL`, this vector will be automatically calculated using the
`x`, `y`, and `z` values in `object`.}

\item{...}{Additional arguments to pass to the primitive creation call. The
available arguments are different for each primitive, and are documented in
the official Blender documentation at
\url{https://docs.blender.org/api/blender_python_api_current/bpy.ops.mesh.html}.}
}
\value{
A length 1 character vector containing the Blender Python script with
code for creating primitives added.
}
\description{
This function generates code that will, when run inside Blender, create mesh
primitives within a scene.
}
