% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{all_missing}
\alias{all_missing}
\title{Check to see if all values in a vector are missing (NULL, NA, INF, or NaN)}
\usage{
all_missing(obj)
}
\arguments{
\item{obj}{The vector of values to check}
}
\description{
Check to see if all values in a vector are missing (NULL, NA, INF, or NaN)
}
\keyword{internal}
