% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{appendix<-}
\alias{appendix<-}
\alias{appendix<-,mvdf_obj-method}
\title{Set appendix values for an `mvdf_obj` object.}
\usage{
appendix(x) <- value

\S4method{appendix}{mvdf_obj}(x) <- value
}
\arguments{
\item{x}{The object to set the appendix within.}

\item{value}{The data to replace the appendix with.}
}
\description{
Set appendix values for an `mvdf_obj` object.
}
