% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame,mvdf_obj-method}
\alias{as.data.frame,mvdf_obj-method}
\title{Coerce to a data frame}
\usage{
\S4method{as.data.frame}{mvdf_obj}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An object inheriting from `mvdf`.}

\item{row.names, optional, ...}{Arguments passed to [base::as.data.frame].}
}
\description{
Coerce any object inheriting from `mvdf` to a data frame.
}
