% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{build_location}
\alias{build_location}
\title{Build the location argument for exporting classes inheriting from `mvdf_obj`}
\usage{
build_location(object)
}
\arguments{
\item{object}{An object inheriting from `mvdf_obj`}
}
\description{
Build the location argument for exporting classes inheriting from `mvdf_obj`
}
\seealso{
Other utilities: 
\code{\link{deg_to_rad}()},
\code{\link{pythonize_booleans}()}
}
\concept{utilities}
\keyword{internal}
