% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blender_frontmatter.R
\name{create_blender_frontmatter}
\alias{create_blender_frontmatter}
\title{Create a Blender rendering script with standard beginning boilerplate}
\usage{
create_blender_frontmatter(
  imports = c("bpy", "mathutils", "math"),
  delete = c("Cube", "Camera", "Light"),
  before = NULL,
  after = NULL
)
}
\arguments{
\item{imports}{A character vector of Python packages to import at the start
of the script. Packages are imported without aliases, so that script-building
functions may always expect to use the full package name. Defaults to
`bpy`, `mathutils`, and `math`. Set to `NULL` to not import any packages.}

\item{delete}{A character vector of objects in the default Blender scene to
delete. Defaults to all objects in the default scene (`Cube`, `Camera`, and
`Light`). Set to `NULL` to not delete any objects.}

\item{before}{Optionally, a character vector containing lines of code to add
before package imports.}

\item{after}{Optionally, a character vector containing lines of code to add
after object deletion.}
}
\value{
A length 1 character vector containing the introductory boilerplate
for a Blender Python script.
}
\description{
Create a Blender rendering script with standard beginning boilerplate
}
