% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{create_options}
\alias{create_options}
\title{Create an options string from a ... object}
\usage{
create_options(dts)
}
\arguments{
\item{dts}{The results of list(...) to create an options string from}
}
\description{
Create an options string from a ... object
}
\keyword{internal}
