% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{eval_arg}
\alias{eval_arg}
\title{Standard way to evaluate arguments for class-creating functions}
\usage{
eval_arg(data, arg)
}
\arguments{
\item{data}{The data frame to try retrieving values from.}

\item{arg}{The argument to evaluate.}
}
\description{
This approach enables users to specify arguments as quoted strings
referencing column names (passes the first try, passes the second try), as
unquoted column names evaluated via rlang's NSE (fails the first try, passes
the second try), or as pure values (passes the first try, fails the second
try).
}
\details{
If the user provides an unquoted string that does not correspond to a column,
this function returns NULL.
}
\keyword{internal}
