% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendix.R
\name{appendix}
\alias{appendix}
\alias{appendix,mvdf_obj-method}
\title{Retrieve the appendix from a `forthetrees` object}
\usage{
appendix(object)

\S4method{appendix}{mvdf_obj}(object)
}
\arguments{
\item{object}{The object to retrieve the appendix from.}
}
\description{
Retrieve the appendix from a `forthetrees` object
}
