% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata,mvdf_obj-method}
\title{Retrieve the metadata data frame from a `forthetrees` object}
\usage{
metadata(object)

\S4method{metadata}{mvdf_obj}(object)
}
\arguments{
\item{object}{The object to retrieve the metadata data frame from.}
}
\description{
Retrieve the metadata data frame from a `forthetrees` object
}
