% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mvdf.R
\name{mvdf}
\alias{mvdf}
\alias{mvdf,mvdf_obj-method}
\title{Retrieve mvdf values from a `mvdf_obj` object.}
\usage{
mvdf(object)

\S4method{mvdf}{mvdf_obj}(object)
}
\arguments{
\item{object}{The `mvdf_obj` object to retrieve the mvdf for.}
}
\description{
Retrieve mvdf values from a `mvdf_obj` object.
}
