% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{headtail}
\alias{headtail}
\alias{tail,mvdf_obj-method}
\title{Return the first or last parts of objects inheriting from `mvdf_obj`}
\usage{
\S4method{head}{mvdf_obj}(x, n = 6L, ...)

\S4method{tail}{mvdf_obj}(x, n = 6L, ...)
}
\arguments{
\item{x}{An object inheriting from `mvdf_obj`}

\item{n}{an integer vector of length up to nrow(mvdf(x)). Values specify the
indices to be selected along the rows of the object. A positive value of n[i]
includes the first/last n[i] indices in that dimension, while a negative
value excludes the last/first abs(n[i]), including all remaining indices.
NA or non-specified values (when length(n) < length(dim(x))) select all
indices in that dimension. Must contain at least one non-missing value.}

\item{...}{Ignored.}
}
\description{
Return the first or last parts of objects inheriting from `mvdf_obj`
}
