% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is_missing}
\alias{is_missing}
\title{Check values in a vector for missingness (NULL, NA, INF, or NaN)}
\usage{
is_missing(obj)
}
\arguments{
\item{obj}{The vector of values to check}
}
\description{
Check values in a vector for missingness (NULL, NA, INF, or NaN)
}
\keyword{internal}
