% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{metadata<-}
\alias{metadata<-}
\alias{metadata<-,mvdf_obj-method}
\title{Set metadata values for an `mvdf_obj` object.}
\usage{
metadata(x) <- value

\S4method{metadata}{mvdf_obj}(x) <- value
}
\arguments{
\item{x}{The object to set the metadata within.}

\item{value}{The data to replace the metadata with.}
}
\description{
Set metadata values for an `mvdf_obj` object.
}
