% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdf-package.R
\docType{package}
\name{mvdf-package}
\alias{mvdf-package}
\alias{_PACKAGE}
\title{mvdf: A Minimum Viable Data Format for 3D Rendering via Blender}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A small, self-contained, minimum viable data format providing a 
    standard interface for using R as a front-end for the Blender 3D rendering
    program. The core approach centers around an S4 class, 'mvdf', with getter,
    setter, and validation methods designed to be extended for more specific 
    rendering approaches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikemahoney218.github.io/mvdf/}
  \item Report bugs at \url{https://github.com/mikemahoney218/mvdf/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
