% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mvdf.R
\name{mvdf_obj}
\alias{mvdf_obj}
\title{Construct a Minimum Viable Data Frame object}
\usage{
mvdf_obj(
  data = NULL,
  x = "x",
  y = "y",
  z = "z",
  idx = "idx",
  metadata = NULL,
  appendix = NULL
)
}
\arguments{
\item{data}{Optionally, a data frame containing all the data necessary to
create a `mvdf_obj`. If left `NULL`, then `x`, `y`, `z`, and `idx` are
interpreted as the values to use for each slot; if not `NULL`, `x`, `y`, `z`,
and `idx` are interpreted as the names of columns in `data` containing the
values for each slot.}

\item{x, y, z}{Numeric: distance of the origin of the object from the origin of
the grid system (the central point at 0, 0, 0) in meters in the given
direction. Must have no `NA`, `NULL`, `NaN`, `Inf`, or `-Inf` values. If
`data` is not `NULL`, the names of columns in `data` with values for the
respective slot.}

\item{idx}{Character: a unique identifier (or "index") for each object to be
modeled. Must be unique with no `NA` or `NULL` values, but otherwise is
not validated. If  `data` is not `NULL`, the names of columns in `data` with
values for the slot. If left `NULL`, a sequential index is generated.}

\item{metadata}{Data frame: a table containing additional information on the
objects to be modeled. Optional, but if this slot is used then the data frame
must contain a column named `idx` which should correspond to the `idx` slot.
Only the existence of this column is validated.}

\item{appendix}{List: additional data produced in the generation of the
object. Not validated; any additional outputs that don't map to modeled
objects may be inserted here.}
}
\description{
Construct a Minimum Viable Data Frame object
}
