% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdf_simple_material.R
\docType{class}
\name{mvdf_simple_material-class}
\alias{mvdf_simple_material-class}
\title{Class to attach basic material data to `mvdf_obj` objects}
\description{
Class to attach basic material data to `mvdf_obj` objects
}
\section{Slots}{

\describe{
\item{\code{x,y,z}}{Numeric: distance of the origin of the object from the origin of
the grid system (the central point at 0, 0, 0) in meters in the given
direction. Must have no `NA`, `NULL`, `NaN`, `Inf`, or `-Inf` values.}

\item{\code{idx}}{Character: a unique identifier (or "index") for each object to be
modeled. Must be unique with no `NA` or `NULL` values, but otherwise is
not validated.}

\item{\code{metadata}}{Data frame: a table containing additional information on the
objects to be modeled. Optional, but if this slot is used then the data frame
must contain a column named `idx` which should correspond to the `idx` slot.
Only the existence of this column is validated.}

\item{\code{appendix}}{List: additional data produced in the generation of the
object. Not validated; any additional outputs that don't map to modeled
objects may be inserted here.}

\item{\code{diffuse_color}}{Diffuse color of the material, as an RGBA array of floats
scaled from 0-1.}

\item{\code{metallic}}{Amount of mirror reflection for raytrace, as a float from
0-1.}

\item{\code{roughness}}{Roughness of the material, as a float from 0-1.}
}}

\seealso{
Other classes and related functions: 
\code{\link{mvdf_obj-class}}
}
\concept{classes and related functions}
