% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{mvdf<-}
\alias{mvdf<-}
\alias{mvdf<-,mvdf_obj-method}
\title{Set mvdf values for an `mvdf_obj` object.}
\usage{
mvdf(x) <- value

\S4method{mvdf}{mvdf_obj}(x) <- value
}
\arguments{
\item{x}{The object to set the mvdf within.}

\item{value}{The data to replace the mvdf with.}
}
\description{
Set mvdf values for an `mvdf_obj` object.
}
