% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{nrow,mvdf_obj-method}
\alias{nrow,mvdf_obj-method}
\alias{ncol,mvdf_obj-method}
\title{The Number of Rows/Columns of an mvdf}
\usage{
\S4method{nrow}{mvdf_obj}(x)

\S4method{ncol}{mvdf_obj}(x)
}
\arguments{
\item{x}{An object inheriting from `mvdf_obj`}
}
\value{
An integer of length 1 or NULL.
}
\description{
`nrow` and `ncol` return the number of rows or columns present in `mvdf(x)`.
}
