% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_methods.R
\name{plot,mvdf_obj-method}
\alias{plot,mvdf_obj-method}
\title{Generic X-Y Plotting}
\usage{
\S4method{plot}{mvdf_obj}(x, y, ...)
}
\arguments{
\item{x}{Any object inheriting from mvdf_obj.}

\item{y, ...}{Arguments passed to base `plot`.}
}
\description{
Draw a basic paired scatter plot from the mvdf of an object inheriting from
`mvdf_obj`.
}
