% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pythonize_booleans}
\alias{pythonize_booleans}
\title{Convert R logical values to Python equivalents}
\usage{
pythonize_booleans(dots)
}
\arguments{
\item{dots}{The results of running `list(...)` in an exporter function.}
}
\description{
Convert R logical values to Python equivalents
}
\seealso{
Other utilities: 
\code{\link{build_location}()},
\code{\link{deg_to_rad}()}
}
\concept{utilities}
\keyword{internal}
