% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{set_appendix}
\alias{set_appendix}
\title{Set appendix values for objects subclassing `mvdf`}
\usage{
set_appendix(appendix, object, mvdf = NULL, metadata = NULL, ...)
}
\arguments{
\item{appendix}{The appendix to include in the new object.}

\item{object}{The object to update.}

\item{mvdf}{The minimum viable data frame required by the S4 class. If `NULL`
(the default), uses the mvdf from `object`.}

\item{metadata}{The metadata to include in the new object. If `NULL`
(the default), uses the metadata from `object`.}

\item{...}{Any additional arguments used in the constructor function being
called.}
}
\description{
This function returns a new object of the same class as `object` with the
same mvdf and metadata slots as `object`, but with a new appendix. Use it as
a convenient, pipe-able way to set appendix values for objects subclassing
`mvdf`.
}
