% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{set_metadata}
\alias{set_metadata}
\title{Set metadata values for objects subclassing `mvdf`}
\usage{
set_metadata(metadata, object, mvdf = NULL, appendix = NULL, ...)
}
\arguments{
\item{metadata}{The metadata to include in the new object.}

\item{object}{The object to update.}

\item{mvdf}{The minimum viable data frame required by the S4 class. If `NULL`
(the default), uses the mvdf from `object`.}

\item{appendix}{The appendix to include in the new object. If `NULL`
(the default), uses the appendix from `object`.}

\item{...}{Any additional arguments used in the constructor function being
called.}
}
\description{
This function returns a new object of the same class as `object` with the
same mvdf and appendix slots as `object`, but with a new metadata. Use it as
a convenient, pipe-able way to set metadata values for objects subclassing
`mvdf`.
}
