% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mvdf.R
\name{set_mvdf}
\alias{set_mvdf}
\title{Set slot values for objects subclassing `mvdf`}
\usage{
set_mvdf(mvdf, object, metadata = NULL, appendix = NULL)
}
\arguments{
\item{mvdf}{The minimum viable data frame required by the S4 class}

\item{object}{The}

\item{metadata}{The metadata to include in the new object. If `NULL`
(the default), uses the metadata from `object`.}

\item{appendix}{The appendix to include in the new object. If `NULL`
(the default), uses the appendix from `object`.}
}
\description{
This function returns a new object of the same class as `object` with the
same metadata and appendix slots as `object`, but with other slots taking
values from `mvdf`. Use it as a convenient, pipe-able way to set slot values
for objects subclassing `mvdf`
}
