% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{set_values}
\alias{set_values}
\title{Set values for objects subclassing `mvdf_obj`}
\usage{
set_values(object, mvdf = NULL, metadata = NULL, appendix = NULL, ...)
}
\arguments{
\item{object}{The object to update.}

\item{mvdf}{The minimum viable data frame required by the S4 class. If `NULL`
(the default), uses the mvdf from `object`.}

\item{metadata}{The metadata to include in the new object. If `NULL`
(the default), uses the metadata from `object`.}

\item{appendix}{The appendix to include in the new object. If `NULL`
(the default), uses the appendix from `object`.}

\item{...}{Any additional arguments used in the constructor function being
called.}
}
\description{
This function returns a new object of the same class as `object` with updated
values. Use it as a convenient, pipe-able way to set values for
objects subclassing `mvdf_obj`.
}
