% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show,mvdf_obj-method}
\alias{show,mvdf_obj-method}
\title{Show an object inheriting from `mvdf_obj`}
\usage{
\S4method{show}{mvdf_obj}(object)
}
\arguments{
\item{object}{Any object inheriting from `mvdf_obj`}
}
\value{
`show` returns an invisible `NULL`.
}
\description{
Display the object, by printing, plotting or whatever suits its class.
}
