% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{deg_to_rad}
\alias{deg_to_rad}
\title{Convert decimal degrees to radians}
\usage{
deg_to_rad(deg)
}
\arguments{
\item{deg}{A vector of values, in decimal degrees, to convert to radians}
}
\value{
A vector of the same length in radians
}
\description{
Convert decimal degrees to radians
}
\seealso{
Other utilities: 
\code{\link{build_location}()},
\code{\link{pythonize_booleans}()}
}
\concept{utilities}
\keyword{internal}
